GH-395 Avoid overflow

+++ plugins/resource_monitor_plugin/include/eosio/resource_monitor_plugin/file_space_handler.hpp
@@ -189,7 +189,7 @@ namespace eosio::resource_monitor {
       bool     output_threshold_warning {true};
 
    private:
-      uint32_t to_gib(uint64_t bytes) {
+      uint64_t to_gib(uint64_t bytes) {
          return bytes/1024/1024/1024;
       }
